+++ src/characters/abstract/Obstacle.ts
@@ -1,7 +1,7 @@
 import Game from "../../Game";
 import { takeDamage } from "../../features/turtleMonitor/turtleReducers";
 import store from "../../store";
-import { generateRandomBit } from "../../utils/generic";
+import { generateRandomBit, vibrate } from "../../utils/generic";
 import IObstacle from "../interfaces/IObstacle";
 import NonMain from "./NonMain";
 
@@ -21,6 +21,7 @@ abstract class Obstacle extends NonMain implements IObstacle {
   handleTurtleCollision(): void {
     store.dispatch(takeDamage({ turtle: { lifeValue: this._damage } }));
     super.handleTurtleCollision();
+    vibrate();
   }
 
   /**