+++ changelog.md
@@ -2,6 +2,7 @@
 
 - Initial cleanup of error dialogs from global state dependency
 - Add desktop notifications
+- Mobile: Add 300ms vibration on collision with obstacles
 
 # 0.7.6.2 Beta (Malta&Gozo hatching Edition) - 27/08/2025 15:00 UTC+1