+++ changelog.md
@@ -1,6 +1,8 @@
-# 0.7.7 Beta (Malta&Gozo hatching Edition) -
+# 0.7.7 Beta -
 
-- Cleanup error dialogs from global state dependency
+- Migrate to vanilla JS and web components
+- Minor UI changes
+- Rename game to Turtle Core
 
 # 0.7.6.2 Beta (Malta&Gozo hatching Edition) - 27/08/2025 15:00 UTC+1